Address ledger PR review findings
