+++ crates/axon-ledger/Cargo.toml
@@ -11,6 +11,7 @@ async-trait = "0.1"
 axon-api = { path = "../axon-api" }
 axon-error = { path = "../axon-error" }
 chrono = "0.4"
+serde = "1"
 serde_json = "1"
 sqlx = { version = "0.8", features = ["runtime-tokio-rustls", "sqlite"] }
 tokio = { version = "1", features = ["macros", "rt", "sync"] }
