+++ docs/reference/api/dto.md
@@ -12,6 +12,6 @@ Generated by `cargo xtask schemas api`.
 | `crates/axon-api/src/source/ids.rs` | `sha256:7021f1ca8e0f68587f4496bb8a0a79d50e0765dbf44ede057e897708d073a004` |
 | `crates/axon-api/src/source/lifecycle.rs` | `sha256:c3a302298345364949c6f63137656e507c3fb2d50f28a4e3945e10de47ec905e` |
 | `crates/axon-api/src/source/stage.rs` | `sha256:118e99fc369f22426f522c1add2f2a1ff1115311eb9472439bd751a3045eb4a7` |
-| `crates/axon-api/src/source/state.rs` | `sha256:51c6bef98b3137ef54573202e03a9199cbd398b6a656272f0c42e6495ac5c01f` |
+| `crates/axon-api/src/source/state.rs` | `sha256:2aa3fc905908f9f8820debed7b3c99eb7334e8bf9c2916f69a2245cb1660dd47` |
 | `crates/axon-api/src/source/status.rs` | `sha256:3313680c7f0c4dfa5ffc3ebb33939c4659bf0b7465de5cdb83e858ee086da6b3` |
 | `crates/axon-error/src/api_error.rs` | `sha256:3f04b3919509ee296cd5badafac5a9dcbc653215827e70f0d8bf98bbb0eb5f7e` |
